requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
URLObject==2.4.3
Werkzeug==3.1.3
WTForms==3.2.1
